Renamed to pathtraversal
