+++ src/main/java/org/owasp/webgoat/lessons/pathtraversal/ProfileUploadFix.java
@@ -1,4 +1,4 @@
-package org.owasp.webgoat.lessons.path_traversal;
+package org.owasp.webgoat.lessons.pathtraversal;
 
 import org.owasp.webgoat.container.assignments.AssignmentHints;
 import org.owasp.webgoat.container.assignments.AttackResult;
@@ -16,23 +16,24 @@ import static org.springframework.http.MediaType.ALL_VALUE;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 
 @RestController
-@AssignmentHints({"path-traversal-profile.hint1", "path-traversal-profile.hint2", "path-traversal-profile.hint3"})
-public class ProfileUpload extends ProfileUploadBase {
+@AssignmentHints({"path-traversal-profile-fix.hint1", "path-traversal-profile-fix.hint2", "path-traversal-profile-fix.hint3"})
+public class ProfileUploadFix extends ProfileUploadBase {
 
-    public ProfileUpload(@Value("${webgoat.server.directory}") String webGoatHomeDirectory, WebSession webSession) {
+    public ProfileUploadFix(@Value("${webgoat.server.directory}") String webGoatHomeDirectory, WebSession webSession) {
         super(webGoatHomeDirectory, webSession);
     }
 
-    @PostMapping(value = "/PathTraversal/profile-upload", consumes = ALL_VALUE, produces = APPLICATION_JSON_VALUE)
+    @PostMapping(value = "/PathTraversal/profile-upload-fix", consumes = ALL_VALUE, produces = APPLICATION_JSON_VALUE)
     @ResponseBody
-    public AttackResult uploadFileHandler(@RequestParam("uploadedFile") MultipartFile file, @RequestParam(value = "fullName", required = false) String fullName) {
-        return super.execute(file, fullName);
+    public AttackResult uploadFileHandler(
+            @RequestParam("uploadedFileFix") MultipartFile file,
+            @RequestParam(value = "fullNameFix", required = false) String fullName) {
+        return super.execute(file, fullName != null ? fullName.replace("../", "") : "");
     }
 
-    @GetMapping("/PathTraversal/profile-picture")
+    @GetMapping("/PathTraversal/profile-picture-fix")
     @ResponseBody
     public ResponseEntity<?> getProfilePicture() {
         return super.getProfilePicture();
     }
-
 }
