+++ src/main/resources/lessons/pathtraversal/css/path_traversal.css
