+++ src/main/resources/lessons/pathtraversal/documentation/PathTraversal_intro.adoc
