+++ src/main/resources/lessons/pathtraversal/documentation/PathTraversal_retrieval.adoc
