+++ src/main/resources/lessons/pathtraversal/documentation/PathTraversal_upload_mitigation.adoc
