+++ src/main/resources/lessons/pathtraversal/documentation/PathTraversal_upload_remove_user_input.adoc
