+++ src/main/resources/lessons/pathtraversal/documentation/PathTraversal_zip_slip_assignment.adoc
