+++ src/main/resources/lessons/pathtraversal/images/cats/2.jpg
