+++ src/main/resources/lessons/pathtraversal/js/path_traversal.js
